move settings to common, fix data path, fix broken imports
